Merge branch 'develop' into python_c_api_update

# Conflicts:
#	examples/fastpointcloud/main.cpp
#	examples/green_screen/main.cpp
#	examples/k4arecord_custom_track/main.c
#	examples/opencv_compatibility/main.cpp
#	examples/streaming/main.c
#	examples/transformation/main.cpp
#	examples/undistort/main.cpp
#	examples/viewer/opengl/main.cpp
#	src/csharp/SDK/ColorModeInfo.cs
#	src/csharp/SDK/DepthModeInfo.cs
#	src/csharp/SDK/DeviceInfo.cs
#	src/csharp/SDK/FPSModeInfo.cs
#	src/csharp/Tests/UnitTests/DeviceFunctionTests.cs
#	src/python/k4a/build_wheel.csh
#	src/python/k4a/build_wheel.ps1
#	src/python/k4a/docs/building.md
#	src/python/k4a/docs/examples.md
#	src/python/k4a/docs/testing.md
#	src/python/k4a/examples/image_transformations.py
#	src/python/k4a/examples/simple_viewer.py
#	src/python/k4a/examples/the_basics.py
#	src/python/k4a/src/k4a/__init__.py
#	src/python/k4a/src/k4a/_bindings/calibration.py
#	src/python/k4a/src/k4a/_bindings/device.py
#	src/python/k4a/src/k4a/_bindings/image.py
#	src/python/k4a/src/k4a/_bindings/k4a.py
#	src/python/k4a/src/k4a/_bindings/k4atypes.py
#	src/python/k4a/src/k4a/_libs/README.md
#	src/python/k4a/tests/test_config.py
#	src/python/k4a/tests/test_functional_api_azurekinect.py
#	src/python/k4a/tests/test_functional_ctypes_azurekinect.py
#	src/python/k4a/tests/test_unit_k4atypes.py
#	tools/k4arecorder/main.cpp
#	tools/k4aviewer/k4adevicedockcontrol.cpp

+++ src/python/k4a/src/k4a/_bindings/image.py
@@ -411,7 +411,7 @@ class Image:
         width_pixels = _ctypes.c_int(width_pixels_custom)
         height_pixels = _ctypes.c_int(height_pixels_custom)
         stride_bytes = _ctypes.c_int(stride_bytes_custom)
-        size_bytes = _ctypes.c_int(size_bytes_custom)
+        size_bytes = _ctypes.c_size_t(size_bytes_custom)
 
         # Use the ndarray sizes if the custom size info is not passed in.
         if width_pixels == 0:
@@ -421,7 +421,7 @@ class Image:
             height_pixels = _ctypes.c_int(arr.shape[1])
 
         if size_bytes == 0:
-            size_bytes = _ctypes.c_int(arr.itemsize * arr.size)
+            size_bytes = _ctypes.c_size_t(arr.itemsize * arr.size)
 
         if len(arr.shape) > 2 and stride_bytes == 0:
             stride_bytes = _ctypes.c_int(arr.shape[2])
